cleanup: new files for components and overall cleaning up in the code
